+++ lib/core/widgets/form_field_widget.dart
@@ -1,5 +1,27 @@
 import 'package:flutter/material.dart';
-import 'package:money/core/widgets/form_field_switch.dart';
+
+Widget myFormField({
+  required final String title,
+  required final Widget child,
+  final bool isReadOnly = false,
+}) {
+  return InputDecorator(
+    decoration: myFormFieldDecoration(fieldName: title, isReadOnly: isReadOnly),
+    child: child,
+  );
+}
+
+InputDecoration myFormFieldDecoration({
+  required final fieldName,
+  required final bool isReadOnly,
+}) {
+  return InputDecoration(
+    labelText: fieldName,
+    contentPadding: isReadOnly ? const EdgeInsets.symmetric(horizontal: 12) : null,
+    // some padding to match the Editable fields that have a border and padding
+    border: const OutlineInputBorder(),
+  );
+}
 
 /// Hybrid widget Text on the left, custom widget on the right
 class MyFormFieldForWidget extends StatefulWidget {
@@ -36,7 +58,7 @@ class MyFormFieldForWidgetState extends State<MyFormFieldForWidget> {
       opacity: widget.isReadOnly ? 0.5 : 1.0,
       child: TextFormField(
         controller: controller,
-        decoration: getFormFieldDecoration(
+        decoration: myFormFieldDecoration(
           fieldName: widget.title,
           isReadOnly: widget.isReadOnly,
         ),