+++ lib/data/storage/import/import_wizard.dart
@@ -1,43 +1,58 @@
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
+import 'package:get/get.dart';
 import 'package:money/core/widgets/dialog/dialog.dart';
 import 'package:money/core/widgets/gaps.dart';
 import 'package:money/core/widgets/wizard_choice.dart';
+import 'package:money/data/storage/import/import_investment.dart';
 import 'package:money/data/storage/import/import_qfx.dart';
 import 'package:money/data/storage/import/import_qif.dart';
 import 'package:money/data/storage/import/import_transactions_from_text.dart';
 
-void showImportTransactionsWizard(
-  final BuildContext context, [
-  String? initialText,
-]) {
+void showImportTransactionsWizard({
+  bool includeInvestment = false,
+}) {
+  final BuildContext context = Get.context!;
+
   adaptiveScreenSizeDialog(
     context: context,
     captionForClose: 'Cancel',
     title: 'Import transactions',
-    child: Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        WizardChoice(
-          title: 'From QFX/QIF file',
-          description: 'Use existing or downloaded files from local device.',
-          onPressed: () {
-            Navigator.of(context).pop(true);
-            onImportFromFile(context);
-          },
-        ),
-        gapHuge(),
-        WizardChoice(
-          title: 'Manual bulk text input',
-          description:
-              'Refer to your online statements, then Copy & Paste text or use OCR to extract the [Dates | Memos | Amounts].',
-          onPressed: () {
-            Navigator.of(context).pop(true);
-            showImportTransactionsFromTextInput(context);
-          },
-        ),
-      ],
+    child: SingleChildScrollView(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        spacing: 40,
+        children: [
+          gapMedium(),
+          WizardChoice(
+            title: 'From QFX/QIF file',
+            description: 'Use existing or downloaded files from local device.',
+            onPressed: () {
+              Navigator.of(context).pop(true);
+              onImportFromFile(context);
+            },
+          ),
+          WizardChoice(
+            title: 'Manual bulk text input',
+            description:
+                'Refer to your online statements, then Copy & Paste text or use OCR to extract the [Dates | Memos | Amounts].',
+            onPressed: () {
+              Navigator.of(context).pop(true);
+              showImportTransactionsFromTextInput(context);
+            },
+          ),
+          if (includeInvestment)
+            WizardChoice(
+              title: 'Investment Transaction',
+              description: 'Buy/Sell/Dividend.',
+              onPressed: () {
+                Navigator.of(context).pop(true);
+                showImportInvestment(context);
+              },
+            ),
+        ],
+      ),
     ),
   );
 }